requirements.txt
asgiref==3.3.1
Authlib==0.15.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.1.6
django-salesforce==3.1
djangorestframework==3.12.2
idna==2.10
pycparser==2.20
pytz==2021.1
requests==2.25.1
simple-salesforce==1.10.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
